224. Basic Calculator 227. Basic Calculator II

+++ leetcode-cpp/basic-calculator-ii/basic-calculator-ii.cpp
@@ -1,5 +1,5 @@
 //
-// Created by smile on 2021/6/20.
+// Created by smile on 2021/9/25.
 //
 
 #include <string>
@@ -102,13 +102,13 @@ private:
 
     // 计算后缀表达式
     double postfixCalculate() {
-        double first, second;
-        std::stack<double> nums;
+        int first, second;
+        std::stack<int> nums;
 
         for (const auto &item: postfix) {
             if (item.Type != OPERATOR_TYPE) {
                 if (item.Type == INTEGER_TYPE) {
-                    nums.emplace((double) item.Value.Integer);
+                    nums.emplace(item.Value.Integer);
                 }
                 if (item.Type == DECIMAL_TYPE) {
                     nums.emplace(item.Value.Decimal);
